change: add e2e tests, fix worfklow with inputs
